deleted HellowWorld Test, added a Test for my usecase
