+++ src/main/java/Use_Case_Interactors/PlotManager.java
@@ -1,7 +1,6 @@
-package Use_Cases;
+package Use_Case_Interactors;
 import Entities.Plot;
 
-import java.lang.reflect.Array;
 import java.util.ArrayList;
 
 public class PlotManager {
