+++ src/main/java/endgameconditions_feature/EndgameConditions_InputBoundary.java
@@ -1,0 +1,4 @@
+package endgameconditions_feature;
+
+public class EndgameConditions_InputBoundary {
+}
