+++ src/main/java/endgameconditions_feature/EndgameConditions_OutputBoundary.java
@@ -1,0 +1,4 @@
+package endgameconditions_feature;
+
+public class EndgameConditions_OutputBoundary {
+}
