+++ src/test/java/EndGameConditions_UseCase_Test.java
@@ -1,0 +1,13 @@
+import endgameconditions_feature.EndgameConditions;
+import org.junit.Test;
+
+public class EndGameConditions_UseCase_Test {
+
+    @Test(timeout = 50)
+    public void testCheckWinMoney()
+    {
+
+        EndgameConditions.checkWinMoney()
+        assertTrue();
+    }
+}
